platform/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic==0.39.0
httpx==0.27.0
openpyxl==3.1.5
pandas==2.2.0
scipy==1.14.0
numpy==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
psycopg2-binary==2.9.9
slowapi==0.1.9
sortedcontainers==2.4.0
sse-starlette==2.1.0
